Refactor quotes hook
