+++ src/hooks/useBestQuote/index.ts
@@ -3,27 +3,12 @@ import { useState } from 'react'
 import { useNetwork } from 'wagmi'
 
 import { BigNumber } from '@ethersproject/bignumber'
-import {
-  getExchangeIssuanceLeveragedQuote,
-  getExchangeIssuanceZeroExQuote,
-  SwapData,
-  ZeroExApi,
-} from '@indexcoop/index-exchange-issuance-sdk'
+import { SwapData, ZeroExApi } from '@indexcoop/index-exchange-issuance-sdk'
 
-import { MAINNET } from 'constants/chains'
 import { IndexApiBaseUrl } from 'constants/server'
-import {
-  eligibleLeveragedExchangeIssuanceTokens,
-  ETH,
-  icETHIndex,
-  IndexToken,
-  JPGIndex,
-  STETH,
-  Token,
-} from 'constants/tokens'
+import { Token } from 'constants/tokens'
 import { useBalances } from 'hooks/useBalance'
 import { toWei } from 'utils'
-import { getExchangeIssuanceGasEstimate } from 'utils/exchangeIssuanceGasEstimate'
 import { getFullCostsInUsd } from 'utils/exchangeIssuanceQuotes'
 import { GasStation } from 'utils/gasStation'
 import { getAddressForToken } from 'utils/tokens'
@@ -33,7 +18,10 @@ import {
   ZeroExData,
 } from 'utils/zeroExUtils'
 
-import { useWallet } from './useWallet'
+import { useWallet } from '../useWallet'
+
+import { getEILeveragedQuote } from './exchangeIssuanceLeveraged'
+import { getEIZeroExQuote } from './exchangeIssuanceZeroEx'
 
 export enum QuoteType {
   notAvailable = 'notAvailable',
@@ -87,65 +75,6 @@ type QuoteResult = {
 
 // To determine if price impact for DEX is smaller 5%
 export const maxPriceImpact = 5
-
-/* Determines if the token is eligible for Leveraged Exchange Issuance */
-const isEligibleLeveragedToken = (token: Token) =>
-  eligibleLeveragedExchangeIssuanceTokens.includes(token)
-
-export function isEligibleTradePairZeroEx(
-  inputToken: Token,
-  outputToken: Token
-): boolean {
-  if (
-    inputToken.symbol === icETHIndex.symbol ||
-    outputToken.symbol === icETHIndex.symbol
-  )
-    return false
-
-  if (
-    inputToken.symbol === IndexToken.symbol ||
-    outputToken.symbol === IndexToken.symbol
-  )
-    return false
-
-  if (
-    inputToken.symbol === JPGIndex.symbol ||
-    outputToken.symbol === JPGIndex.symbol
-  )
-    // temporarily - disabled JPG for EI0x
-    return false
-
-  return true
-}
-
-/* Determines if the token pair is eligible for Leveraged Exchange Issuance */
-export const isEligibleTradePair = (
-  inputToken: Token,
-  outputToken: Token,
-  isIssuance: boolean
-) => {
-  const tokenEligible = isIssuance
-    ? isEligibleLeveragedToken(outputToken)
-    : isEligibleLeveragedToken(inputToken)
-
-  const isIcEth =
-    inputToken.symbol === icETHIndex.symbol ||
-    outputToken.symbol === icETHIndex.symbol
-
-  if (tokenEligible && isIcEth && isIssuance) {
-    // Only ETH or stETH is allowed as input for icETH issuance at the moment
-    return (
-      inputToken.symbol === ETH.symbol || inputToken.symbol === STETH.symbol
-    )
-  }
-
-  if (tokenEligible && isIcEth && !isIssuance) {
-    // Only ETH is allowed as output for icETH redeeming at the moment
-    return outputToken.symbol === ETH.symbol
-  }
-
-  return tokenEligible
-}
 
 export function getBestQuote(
   fullCosts0x: number | null,
@@ -231,7 +160,7 @@ const defaultQuoteResult: QuoteResult = {
   success: false,
 }
 
-export const useBestTradeOption = () => {
+export const useBestQuote = () => {
   const { provider } = useWallet()
   const { chain } = useNetwork()
   const { getBalance } = useBalances()
@@ -323,7 +252,7 @@ export const useBestTradeOption = () => {
         }
       : null
 
-    /* Determine set token amount based on different factors */
+    /* Determine Set token amount based on different factors */
     let setTokenAmount = getSetTokenAmount(
       isIssuance,
       sellTokenAmount,
@@ -336,11 +265,6 @@ export const useBestTradeOption = () => {
     const gasStation = new GasStation(provider)
     const gasPrice = await gasStation.getGasPrice()
 
-    /* Check for Exchange Issuance option */
-    let exchangeIssuanceZeroExQuote: ExchangeIssuanceZeroExQuote | null = null
-    let exchangeIssuanceLeveragedQuote: ExchangeIssuanceLeveragedQuote | null =
-      null
-
     // Create an instance of ZeroExApi (to pass to quote functions)
     const affilliateAddress = '0x37e6365d4f6aE378467b0e24c9065Ce5f06D70bF'
     const networkKey = getNetworkKey(chainId)
@@ -351,118 +275,41 @@ export const useBestTradeOption = () => {
       swapPathOverride
     )
 
-    const inputToken = {
-      symbol: sellToken.symbol,
-      decimals: sellToken.decimals,
-      address: inputTokenAddress,
-    }
-    const outputToken = {
-      symbol: buyToken.symbol,
-      decimals: buyToken.decimals,
-      address: outputTokenAddress,
-    }
+    const exchangeIssuanceLeveragedQuote: ExchangeIssuanceLeveragedQuote | null =
+      await getEILeveragedQuote(
+        isIssuance,
+        inputTokenAddress,
+        outputTokenAddress,
+        sellToken,
+        buyToken,
+        setTokenAmount,
+        sellTokenPrice,
+        nativeTokenPrice,
+        gasPrice,
+        slippage,
+        chainId,
+        provider,
+        zeroExApi
+      )
 
-    const tokenEligibleForLeveragedEI = isEligibleTradePair(
-      sellToken,
-      buyToken,
-      isIssuance
-    )
-    if (tokenEligibleForLeveragedEI) {
-      try {
-        const quoteLeveraged = await getExchangeIssuanceLeveragedQuote(
-          inputToken,
-          outputToken,
-          setTokenAmount,
-          isIssuance,
-          slippage,
-          zeroExApi,
-          provider,
-          chainId ?? 1
-        )
-        if (quoteLeveraged) {
-          const gasLimit = BigNumber.from(1800000)
-          exchangeIssuanceLeveragedQuote = {
-            type: QuoteType.exchangeIssuanceLeveraged,
-            isIssuance,
-            inputToken: sellToken,
-            outputToken: buyToken,
-            gas: gasLimit,
-            gasPrice,
-            gasCosts: gasLimit.mul(gasPrice),
-            fullCostsInUsd: getFullCostsInUsd(
-              quoteLeveraged.inputOutputTokenAmount,
-              gasLimit.mul(gasPrice),
-              sellToken.decimals,
-              sellTokenPrice,
-              nativeTokenPrice
-            ),
-            priceImpact: 0,
-            setTokenAmount,
-            inputOutputTokenAmount: quoteLeveraged.inputOutputTokenAmount,
-            // type specific properties
-            swapDataDebtCollateral: quoteLeveraged.swapDataDebtCollateral,
-            swapDataPaymentToken: quoteLeveraged.swapDataPaymentToken,
-          }
-        }
-        console.log(slippage, slippagePercentage, quoteLeveraged)
-      } catch (e) {
-        console.warn('error when generating leveraged ei option', e)
-      }
-    } else {
-      // For now only allow trade on mainnet, some tokens are disabled
-      const isEligibleTradePair = isEligibleTradePairZeroEx(sellToken, buyToken)
-      if (chainId === MAINNET.chainId && isEligibleTradePair)
-        try {
-          const spendingTokenBalance: BigNumber =
-            getBalance(sellToken.symbol) || BigNumber.from(0)
-          const quote0x = await getExchangeIssuanceZeroExQuote(
-            inputToken,
-            outputToken,
-            setTokenAmount,
-            isIssuance,
-            slippage,
-            zeroExApi,
-            provider,
-            chainId
-          )
-          if (quote0x) {
-            const gasEstimate = await getExchangeIssuanceGasEstimate(
-              provider,
-              chainId,
-              isIssuance,
-              sellToken,
-              buyToken,
-              setTokenAmount,
-              quote0x.inputOutputTokenAmount,
-              spendingTokenBalance,
-              quote0x.componentQuotes
-            )
-            exchangeIssuanceZeroExQuote = {
-              type: QuoteType.exchangeIssuanceZeroEx,
-              isIssuance,
-              inputToken: sellToken,
-              outputToken: buyToken,
-              gas: gasEstimate,
-              gasPrice,
-              gasCosts: gasEstimate.mul(gasPrice),
-              fullCostsInUsd: getFullCostsInUsd(
-                quote0x.inputOutputTokenAmount,
-                gasEstimate.mul(gasPrice),
-                sellToken.decimals,
-                sellTokenPrice,
-                nativeTokenPrice
-              ),
-              priceImpact: 0,
-              setTokenAmount,
-              inputOutputTokenAmount: quote0x.inputOutputTokenAmount,
-              // type specific properties
-              componentQuotes: quote0x.componentQuotes,
-            }
-          }
-        } catch (e) {
-          console.warn('error when generating zeroexei option', e)
-        }
-    }
+    const inputTokenBalance = getBalance(sellToken.symbol) ?? BigNumber.from(0)
+    const exchangeIssuanceZeroExQuote: ExchangeIssuanceZeroExQuote | null =
+      await getEIZeroExQuote(
+        isIssuance,
+        inputTokenAddress,
+        outputTokenAddress,
+        inputTokenBalance,
+        sellToken,
+        buyToken,
+        setTokenAmount,
+        sellTokenPrice,
+        nativeTokenPrice,
+        gasPrice,
+        slippage,
+        chainId,
+        provider,
+        zeroExApi
+      )
 
     console.log('////////')
     console.log('exchangeIssuanceZeroExQuote', exchangeIssuanceZeroExQuote)
